requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
ci-info==0.1.1
click==7.1.1
colorama==0.4.3
configparser==5.0.0
decorator==4.4.2
etelemetry==0.2.0
filelock==3.0.12
firebase-admin==4.0.1
Flask==1.1.1
future==0.18.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
httplib2==0.17.2
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
more-itertools==8.2.0
msgpack==1.0.0
networkx==2.4
neurdflib==5.0.1
numpy==1.18.2
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
protobuf==3.11.3
prov==1.5.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pydotplus==2.0.2
PyMuPDF==1.16.16
pyparsing==2.4.7
pytest==5.3.2
python-dateutil==2.8.1
pytz==2019.3
pyxnat==1.2.1.0.post3
rdflib==4.2.2
requests==2.23.0
rsa==4.0
scipy==1.4.1
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
wget==3.2
zipp==3.1.0
